Transport crud optimization
The transport crud view was very slow and very unintuitive, lists and maps where passed to the view directly, which made it really slow and hard to mantain. Now, only the needed info is passed to the view and times have significantly improved.

Files have been refactored from TemplateCrud to TransportCrud, which is the right way to name it.